requirements.txt
anthropic>=0.40.0
requests>=2.32.0
gspread>=6.1.0
google-auth>=2.30.0
python-dotenv>=1.0.1
solapi>=5.0.0
tenacity>=9.0.0
